Refactor tests for database implementations into a common file
