+++ backend/package.json
@@ -21,11 +21,13 @@
         "@types/express": "^4.17.2",
         "@types/express-pino-logger": "^4.0.2",
         "@types/jest": "^24.9.1",
+        "@types/jest-each": "^24.3.0",
         "@types/lodash": "^4.14.136",
         "@types/node": "^12.6.9",
         "@types/pino": "^5.15.1",
         "@types/uuid": "^3.4.6",
         "jest": "^26.0.1",
+        "jest-each": "^26.1.0",
         "nodemon": "^2.0.2",
         "ts-jest": "^26.0.0",
         "ts-node": "^8.5.4",
